kvm: separate networking code
