+++ networking/networking.go
@@ -15,14 +15,12 @@
 package networking
 
 import (
-	"encoding/json"
 	"fmt"
 	"log"
 	"net"
 	"os"
 	"syscall"
 
-	"github.com/coreos/rkt/Godeps/_workspace/src/github.com/appc/cni/pkg/ip"
 	"github.com/coreos/rkt/Godeps/_workspace/src/github.com/appc/cni/pkg/ns"
 	"github.com/coreos/rkt/Godeps/_workspace/src/github.com/appc/cni/pkg/plugin"
 	"github.com/coreos/rkt/Godeps/_workspace/src/github.com/appc/spec/schema/types"
@@ -30,7 +28,6 @@ import (
 
 	"github.com/coreos/rkt/common"
 	"github.com/coreos/rkt/networking/netinfo"
-	"github.com/coreos/rkt/networking/tuntap"
 )
 
 const (
@@ -60,111 +57,13 @@ type NetConf struct {
 	MTU    int  `json:"mtu"`
 }
 
-func setupTapDevice() (string, *netlink.Link, error) {
-	ifName, err := tuntap.CreatePersistentIface(tuntap.Tap)
-	if err != nil {
-		return "", nil, fmt.Errorf("tuntap persist %v", err)
-	}
-	var link netlink.Link
-	link, err = netlink.LinkByName(ifName)
-	if err != nil {
-		return "", nil, fmt.Errorf("cannot find link %q: %v", ifName, err)
-	}
-	err = netlink.LinkSetUp(link)
-	if err != nil {
-		return "", nil, fmt.Errorf("cannot set link up %q: %v", ifName, err)
-	}
-	return ifName, &link, nil
-}
-
-func kvmSetupNetAddressing(network *Networking, n activeNet, ifName string) error {
-	// TODO: very ugly hack, that go through upper plugin, down to ipam plugin
-	n.conf.Type = n.conf.IPAM.Type
-	output, err := network.execNetPlugin("ADD", &n, ifName)
-	if err != nil {
-		return fmt.Errorf("problem executing network plugin %q (%q): %v", n.conf.Type, ifName, err)
-	}
-
-	result := plugin.Result{}
-	if err = json.Unmarshal(output, &result); err != nil {
-		return fmt.Errorf("error parsing %q result: %v", n.conf.Name, err)
-	}
-
-	if result.IP4 == nil {
-		return fmt.Errorf("net-plugin returned no IPv4 configuration")
-	}
-
-	n.runtime.IP, n.runtime.Mask, n.runtime.HostIP = result.IP4.IP.IP, net.IP(result.IP4.IP.Mask), result.IP4.Gateway
-	return nil
-}
-
-func KvmSetup(podRoot string, podID types.UUID, fps []ForwardedPort, privateNetList common.PrivateNetList, localConfig string) (*Networking, error) {
-	network := Networking{
-		podEnv: podEnv{
-			podRoot:      podRoot,
-			podID:        podID,
-			netsLoadList: privateNetList,
-			localConfig:  localConfig,
-		},
-	}
-	var e error
-	network.nets, e = network.loadNets()
-	if e != nil {
-		return nil, fmt.Errorf("error loading network definitions: %v", e)
-	}
-
-	for _, n := range network.nets {
-		switch n.conf.Type {
-		case "ptp":
-			ifName, link, err := setupTapDevice()
-			if err != nil {
-				return nil, err
-			}
-			n.runtime.IfName = ifName
-
-			err = kvmSetupNetAddressing(&network, n, ifName)
-			if err != nil {
-				return nil, err
-			}
-
-			// add address to host tap device
-			err = netlink.AddrAdd(
-				*link,
-				&netlink.Addr{
-					IPNet: &net.IPNet{
-						IP:   n.runtime.HostIP,
-						Mask: net.IPMask(n.runtime.Mask),
-					},
-					Label: ifName,
-				})
-			if err != nil {
-				return nil, fmt.Errorf("cannot add address to host tap device %q: %v", ifName, err)
-			}
-
-			if n.conf.IPMasq {
-				chain := "CNI-" + n.conf.Name
-				if err = ip.SetupIPMasq(&net.IPNet{
-					IP:   n.runtime.IP,
-					Mask: net.IPMask(n.runtime.Mask),
-				}, chain); err != nil {
-					return nil, err
-				}
-			}
-		default:
-			return nil, fmt.Errorf("network %q have unsupported type: %q", n.conf.Name, n.conf.Type)
-		}
-	}
-	err := network.forwardPorts(fps, network.GetDefaultIP())
-	if err != nil {
-		return nil, err
-	}
-
-	return &network, nil
-}
-
 // Setup creates a new networking namespace and executes network plugins to
 // setup private networking. It returns in the new pod namespace
-func Setup(podRoot string, podID types.UUID, fps []ForwardedPort, privateNetList common.PrivateNetList, localConfig string) (*Networking, error) {
+func Setup(podRoot string, podID types.UUID, fps []ForwardedPort, privateNetList common.PrivateNetList, localConfig, flavor string) (*Networking, error) {
+	if flavor == "kvm" {
+		return kvmSetup(podRoot, podID, fps, privateNetList, localConfig)
+	}
+
 	// TODO(jonboulle): currently podRoot is _always_ ".", and behaviour in other
 	// circumstances is untested. This should be cleaned up.
 	n := Networking{
@@ -304,23 +203,6 @@ func (n *Networking) Teardown(flavor string) {
 	}
 }
 
-func (e *Networking) teardownKvmNets(nets []activeNet) {
-	for _, n := range nets {
-		switch n.conf.Type {
-		case "ptp":
-			tuntap.RemovePersistentIface(n.runtime.IfName, tuntap.Tap)
-			n.conf.Type = n.conf.IPAM.Type
-
-			_, err := e.execNetPlugin("DEL", &n, n.runtime.IfName)
-			if err != nil {
-				log.Printf("Error executing network plugin: %q", err)
-			}
-		default:
-			log.Printf("Unsupported network type: %q", n.conf.Type)
-		}
-	}
-}
-
 // sets up new netns with just lo
 func basicNetNS() (hostNS, podNS *os.File, err error) {
 	hostNS, podNS, err = newNetNS()
@@ -353,43 +235,6 @@ func (e *Networking) Save() error {
 	}
 
 	return netinfo.Save(e.podRoot, nis)
-}
-
-// NetParams exposes conf(NetConf)/runtime(NetInfo) data to stage1/init client
-type NetParams struct {
-	// runtime based information
-	HostIP  net.IP
-	GuestIP net.IP
-	Mask    net.IP
-	IfName  string
-	// TODO: required for other type of plugins, not yet available because what networking.Networking stores
-	// Net net.IPNet
-
-	// configuration based information
-	Name   string
-	Type   string
-	IPMasq bool
-}
-
-// GetNetworkParameters returns network parameters created
-// by plugins, which are required for stage1 executor to run (only for KVM)
-func (e *Networking) GetNetworkParameters() []NetParams {
-	np := []NetParams{}
-	_ = np
-	for _, an := range e.nets {
-		np = append(np, NetParams{
-			HostIP:  an.runtime.HostIP,
-			GuestIP: an.runtime.IP,
-			IfName:  an.runtime.IfName,
-			Mask:    an.runtime.Mask,
-			// Net: // TODO: from where
-			Name:   an.conf.Name,
-			Type:   an.conf.Type,
-			IPMasq: an.conf.IPMasq,
-		})
-	}
-
-	return np
 }
 
 func newNetNS() (hostNS, childNS *os.File, err error) {
