+++ stage1/init/init.go
@@ -235,28 +235,6 @@ func installAssets() error {
 	return proj2aci.PrepareAssets(assets, "./stage1/rootfs/", nil)
 }
 
-// getKVMNetParams returns additionall parameters that need to be passed to kernel and lkvm tool to configure networks properly
-func getKVMNetParams(network networking.Networking) ([]string, []string, error) {
-	lkvmParams := []string{}
-	kernelParams := []string{}
-
-	for i, netParams := range network.GetNetworkParameters() {
-		//
-
-		// https://www.kernel.org/doc/Documentation/filesystems/nfs/nfsroot.txt
-		// ip=<client-ip>:<server-ip>:<gw-ip>:<netmask>:<hostname>:<device>:<autoconf>:<dns0-ip>:<dns1-ip>
-		var gw string
-		if netParams.IPMasq {
-			gw = netParams.HostIP.String()
-		}
-		kernelParams = append(kernelParams, "ip="+netParams.GuestIP.String()+"::"+gw+":"+netParams.Mask.String()+":"+fmt.Sprintf(networking.IfNamePattern, i)+":::")
-
-		lkvmParams = append(lkvmParams, "--network", "mode=tap,tapif="+netParams.IfName+",host_ip="+netParams.HostIP.String()+",guest_ip="+netParams.GuestIP.String())
-	}
-
-	return lkvmParams, kernelParams, nil
-}
-
 // getArgsEnv returns the nspawn args and env according to the usr used
 func getArgsEnv(p *Pod, flavor string, debug bool, n networking.Networking) ([]string, []string, error) {
 	args := []string{}
@@ -268,7 +246,7 @@ func getArgsEnv(p *Pod, flavor string, debug bool, n networking.Networking) ([]s
 		// TODO: move to path.go
 		kernelPath := filepath.Join(common.Stage1RootfsPath(p.Root), "bzImage")
 		lkvmPath := filepath.Join(common.Stage1RootfsPath(p.Root), "lkvm")
-		lkvmNetParams, kernelNetParams, err := getKVMNetParams(n)
+		lkvmNetParams, kernelNetParams, err := kvm.GetKVMNetParams(n)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -539,11 +517,7 @@ func stage1() int {
 			return 6
 		}
 
-		if flavor == "kvm" {
-			n, err = networking.KvmSetup(root, p.UUID, fps, privNet, localConfig)
-		} else {
-			n, err = networking.Setup(root, p.UUID, fps, privNet, localConfig)
-		}
+		n, err = networking.Setup(root, p.UUID, fps, privNet, localConfig, flavor)
 		if err != nil {
 			fmt.Fprintf(os.Stderr, "Failed to setup network: %v\n", err)
 			return 6
